pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = ["Antti Virtanen <antti.b.virtanen@helsinki.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^2.2.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
coverage = "^7.4.4"
pylint = "^3.1.0"
autopep8 = "^2.1.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
